# docs/swagger.yml
openapi: 3.0.3
info:
  title: Order Management System API
  version: 1.0.0
  description: API para gestionar pedidos, productos, clientes, repartidores y autenticación.
servers:
  - url: http://localhost:3000
    description: Servidor local
  - url: https://order-management-system-995e.onrender.com
    description: Servidor de producción
...
paths:
...
  /api/couriers:
    post:
      tags:
        - Couriers
      summary: Crear courier
      description: Crea un repartidor asociado al usuario autenticado. El nombre debe ser único dentro del
        usuario.
      operationId: createCourier
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourierRequest"
            examples:
              createCourier:
                summary: Crear courier
                value:
                  name: Repartidor 1
                  phone: "1234567890"
                  note: Disponible en horario comercial
      responses:
        "201":
          description: Courier creado correctamente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCourierSuccessResponse"
              examples:
                created:
                  summary: Courier creado
                  value:
                    success: true
                    message: Courier creado correctamente
                    courier:
                      id_courier: 1
                      name: Repartidor 1
                      phone: "1234567890"
                      note: Disponible en horario comercial
        "400":
          description: Error de validación del cuerpo de la petición o JSON inválido.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ValidationErrorResponse"
                  - $ref: "#/components/schemas/GenericErrorResponse"
              examples:
                invalidName:
                  summary: Nombre inválido
                  value:
                    success: false
                    errors:
                      - field: name
                        message: El campo name no puede estar vacío
                invalidPhone:
                  summary: Teléfono inválido
                  value:
                    success: false
                    errors:
                      - field: phone
                        message: El campo phone no es un número de teléfono válido
                malformedJson:
                  summary: JSON inválido
                  value:
                    success: false
                    errors:
                      - message: JSON inválido o mal formado
        "401":
          description: No se envió un access token válido o el encabezado Authorization es inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericErrorResponse"
              examples:
                missingToken:
                  summary: Token requerido
                  value:
                    success: false
                    errors:
                      - message: Token requerido
                invalidHeader:
                  summary: Encabezado inválido
                  value:
                    success: false
                    errors:
                      - message: Encabezado Authorization inválido
                invalidToken:
                  summary: Token inválido
                  value:
                    success: false
                    errors:
                      - message: Token inválido
        "409":
          description: El name ya existe para un courier del usuario autenticado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericErrorResponse"
              examples:
                conflict:
                  summary: Conflicto
                  value:
                    success: false
                    errors:
                      - message: El name ya esta en uso, debe ser unico por courier
        "429":
          description: Se superó el límite de solicitudes de courier.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericErrorResponse"
              examples:
                rateLimit:
                  summary: Límite de solicitudes superado
                  value:
                    success: false
                    errors:
                      - message: Demasiadas peticiones de courier. Intenta nuevamente más tarde.
    get:
      tags:
        - Couriers
      summary: Obtener couriers
      description: Obtiene los couriers activos asociados al usuario autenticado. Permite paginar la
        respuesta con los query params page y limit.
      operationId: getCouriers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: Número de página a consultar. Si no se envía, el servicio utiliza la página 1.
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: limit
          in: query
          required: false
          description: Cantidad de couriers por página. Si no se envía, el servicio utiliza 20. El máximo
            permitido es 50.
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
            example: 20
      responses:
        "200":
          description: Couriers obtenidos correctamente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCouriersSuccessResponse"
              examples:
                withCouriers:
                  summary: Couriers encontrados
                  value:
                    success: true
                    couriers:
                      - id_courier: 1
                        name: Repartidor 1
                        phone: "1234567890"
                        note: Disponible en horario comercial
                      - id_courier: 2
                        name: Repartidor 2
                        phone: null
                        note: null
                empty:
                  summary: Sin couriers
                  value:
                    success: true
                    couriers: []
        "400":
          description: Error de validación de los query params.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
              examples:
                invalidPage:
                  summary: Página inválida
                  value:
                    success: false
                    errors:
                      - field: page
                        message: La página debe ser un número entero
                invalidLimit:
                  summary: Límite inválido
                  value:
                    success: false
                    errors:
                      - field: limit
                        message: El límite debe ser un número entero
        "401":
          description: No se envió un access token válido o el encabezado Authorization es inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericErrorResponse"
              examples:
                missingToken:
                  summary: Token requerido
                  value:
                    success: false
                    errors:
                      - message: Token requerido
                invalidHeader:
                  summary: Encabezado inválido
                  value:
                    success: false
                    errors:
                      - message: Encabezado Authorization inválido
                invalidToken:
                  summary: Token inválido
                  value:
                    success: false
                    errors:
                      - message: Token inválido
        "429":
          description: Se superó el límite de solicitudes de courier.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericErrorResponse"
              examples:
                rateLimit:
                  summary: Límite de solicitudes superado
                  value:
                    success: false
                    errors:
                      - message: Demasiadas peticiones de courier. Intenta nuevamente más tarde.
  /api/couriers/{id}:
    get:
      tags:
        - Couriers
      summary: Obtener courier por identificador
      description: Obtiene un courier del usuario autenticado usando su identificador. Si el courier no
        existe o pertenece a otro usuario, devuelve Courier no encontrado.
      operationId: getCourierById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador del courier. Debe ser un número entero mayor o igual a 1.
          schema:
            type: integer
            minimum: 1
            example: 1
      responses:
        "200":
          description: Courier obtenido correctamente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCourierByIdSuccessResponse"
              examples:
                found:
                  summary: Courier encontrado
                  value:
                    success: true
                    courier:
                      id_courier: 1
                      name: Repartidor 1
                      phone: "1234567890"
                      note: Disponible en horario comercial
        "400":
          description: Error de validación del parámetro id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
              examples:
                invalidId:
                  summary: Identificador inválido
                  value:
                    success: false
                    errors:
                      - field: id
                        message: El identificador debe ser un número entero
                invalidIdMin:
                  summary: Identificador menor a uno
                  value:
                    success: false
                    errors:
                      - field: id
                        message: El identificador debe ser mayor o igual a 1
        "401":
          description: No se envió un access token válido o el encabezado Authorization es inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericErrorResponse"
              examples:
                missingToken:
                  summary: Token requerido
                  value:
                    success: false
                    errors:
                      - message: Token requerido
                invalidHeader:
                  summary: Encabezado inválido
                  value:
                    success: false
                    errors:
                      - message: Encabezado Authorization inválido
                invalidToken:
                  summary: Token inválido
                  value:
                    success: false
                    errors:
                      - message: Token inválido
        "404":
          description: El courier no existe o no pertenece al usuario autenticado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericErrorResponse"
              examples:
                notFound:
                  summary: Recurso no encontrado
                  value:
                    success: false
                    errors:
                      - message: Courier no encontrado
        "429":
          description: Se superó el límite de solicitudes de courier.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericErrorResponse"
              examples:
                rateLimit:
                  summary: Límite de solicitudes superado
                  value:
                    success: false
                    errors:
                      - message: Demasiadas peticiones de courier. Intenta nuevamente más tarde.
    patch:
      tags:
        - Couriers
      summary: Actualizar courier por identificador
      description: Actualiza un courier del usuario autenticado usando su identificador. Devuelve el
        courier actualizado.
      operationId: updateCourierById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador del courier. Debe ser un número entero mayor o igual a 1.
          schema:
            type: integer
            minimum: 1
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourierRequest"
            examples:
              updateCourier:
                summary: Actualizar courier
                value:
                  name: Repartidor actualizado
                  phone: "0987654321"
                  note: Cambio de turno
      responses:
        "200":
          description: Courier actualizado correctamente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateCourierSuccessResponse"
              examples:
                updated:
                  summary: Actualización exitosa
                  value:
                    success: true
                    message: Courier actualizado correctamente
                    courier:
                      id_courier: 1
                      name: Repartidor actualizado
                      phone: "0987654321"
                      note: Cambio de turno
        "400":
          description: Error de validación del parámetro id o JSON inválido.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ValidationErrorResponse"
                  - $ref: "#/components/schemas/GenericErrorResponse"
              examples:
                invalidId:
                  summary: Identificador inválido
                  value:
                    success: false
                    errors:
                      - field: id
                        message: El identificador debe ser un número entero
                invalidIdMin:
                  summary: Identificador menor a uno
                  value:
                    success: false
                    errors:
                      - field: id
                        message: El identificador debe ser mayor o igual a 1
                malformedJson:
                  summary: JSON inválido
                  value:
                    success: false
                    errors:
                      - message: JSON inválido o mal formado
        "401":
          description: No se envió un access token válido o el encabezado Authorization es inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericErrorResponse"
              examples:
                missingToken:
                  summary: Token requerido
                  value:
                    success: false
                    errors:
                      - message: Token requerido
                invalidHeader:
                  summary: Encabezado inválido
                  value:
                    success: false
                    errors:
                      - message: Encabezado Authorization inválido
                invalidToken:
                  summary: Token inválido
                  value:
                    success: false
                    errors:
                      - message: Token inválido
        "404":
          description: El courier no existe o no pertenece al usuario autenticado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericErrorResponse"
              examples:
                notFound:
                  summary: Recurso no encontrado
                  value:
                    success: false
                    errors:
                      - message: Courier no encontrado
        "429":
          description: Se superó el límite de solicitudes de courier.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericErrorResponse"
              examples:
                rateLimit:
                  summary: Límite de solicitudes superado
                  value:
                    success: false
                    errors:
                      - message: Demasiadas peticiones de courier. Intenta nuevamente más tarde.
    delete:
      tags:
        - Couriers
      summary: Eliminar courier por identificador
      description: Marca como eliminado lógicamente un courier del usuario autenticado usando su
        identificador. Devuelve una respuesta 204 sin contenido si la operación es exitosa.
      operationId: deleteCourierById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador del courier. Debe ser un número entero mayor o igual a 1.
          schema:
            type: integer
            minimum: 1
            example: 1
      responses:
        "204":
          description: Courier eliminado correctamente. Sin contenido en la respuesta.
        "400":
          description: Error de validación del path param id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
              examples:
                invalidId:
                  summary: ID inválido
                  value:
                    success: false
                    errors:
                      - field: id
                        message: El identificador debe ser un número entero
                invalidIdMin:
                  summary: ID menor a uno
                  value:
                    success: false
                    errors:
                      - field: id
                        message: El identificador debe ser mayor o igual a 1
        "401":
          description: No se envió un access token válido o el encabezado Authorization es inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericErrorResponse"
              examples:
                missingToken:
                  summary: Token requerido
                  value:
                    success: false
                    errors:
                      - message: Token requerido
                invalidHeader:
                  summary: Encabezado inválido
                  value:
                    success: false
                    errors:
                      - message: Encabezado Authorization inválido
                invalidToken:
                  summary: Token inválido
                  value:
                    success: false
                    errors:
                      - message: Token inválido
        "404":
          description: El courier no existe o no pertenece al usuario autenticado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericErrorResponse"
              examples:
                notFound:
                  summary: Recurso no encontrado
                  value:
                    success: false
                    errors:
                      - message: Courier no encontrado
        "429":
          description: Se superó el límite de solicitudes de courier.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericErrorResponse"
              examples:
                rateLimit:
                  summary: Límite de solicitudes superado
                  value:
                    success: false
                    errors:
                      - message: Demasiadas peticiones de courier. Intenta nuevamente más tarde.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Courier:
      type: object
      required:
        - id_courier
        - name
      properties:
        id_courier:
          type: integer
          format: int32
          example: 1
        name:
          type: string
          example: Repartidor 1
        phone:
          type: string
          nullable: true
          example: "1234567890"
        note:
          type: string
          nullable: true
          example: Disponible en horario comercial
    CreateCourierRequest:
      type: object
      required:
        - name
      additionalProperties: true
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          description: Nombre del repartidor.
          example: Repartidor 1
        phone:
          type: string
          description: Número de teléfono del repartidor.
          example: "1234567890"
        note:
          type: string
          maxLength: 500
          description: Nota opcional sobre el repartidor.
          example: Disponible en horario comercial
    UpdateCourierRequest:
      type: object
      additionalProperties: true
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          description: Nuevo nombre del repartidor.
          example: Repartidor actualizado
        phone:
          type: string
          description: Nuevo número de teléfono del repartidor.
          example: "0987654321"
        note:
          type: string
          maxLength: 500
          description: Nueva nota del repartidor.
          example: Cambio de turno
    CreateCourierSuccessResponse:
      type: object
      required:
        - success
        - message
        - courier
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Courier creado correctamente
        courier:
          $ref: "#/components/schemas/Courier"
    GetCouriersSuccessResponse:
      type: object
      required:
        - success
        - couriers
      properties:
        success:
          type: boolean
          example: true
        couriers:
          type: array
          items:
            $ref: "#/components/schemas/Courier"
    GetCourierByIdSuccessResponse:
      type: object
      required:
        - success
        - courier
      properties:
        success:
          type: boolean
          example: true
        courier:
          $ref: "#/components/schemas/Courier"
    UpdateCourierSuccessResponse:
      type: object
      required:
        - success
        - message
        - courier
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Courier actualizado correctamente
        courier:
          $ref: "#/components/schemas/Courier"
...
    ValidationErrorResponse:
      type: object
      required:
        - success
        - errors
      properties:
        success:
          type: boolean
          example: false
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
    GenericErrorResponse:
      type: object
      required:
        - success
        - errors
      properties:
        success:
          type: boolean
          example: false
        errors:
          type: array
          items:
            $ref: "#/components/schemas/MessageError"
    FieldError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: El email es obligatorio
    MessageError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Token requerido